requirement.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
h11==0.12.0
idna==3.3
mysql-connector-python==8.0.27
protobuf==3.19.1
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.16.0
